+++ public/pages/PreviewDetector/components/FeatureChart/FeatureChart.tsx
@@ -19,19 +19,24 @@ import {
   LineSeries,
   RectAnnotation,
   niceTimeFormatter,
+  Position,
   Settings,
 } from '@elastic/charts';
-import { EuiButton, EuiEmptyPrompt, EuiText } from '@elastic/eui';
-import React from 'react';
+import { EuiEmptyPrompt, EuiText, EuiLink } from '@elastic/eui';
+import React, { useState } from 'react';
 import ContentPanel from '../../../../components/ContentPanel/ContentPanel';
 import { useDelayedLoader } from '../../../../hooks/useDelayedLoader';
 import { Moment } from 'moment';
-import { FeatureAggregationData } from 'public/models/interfaces';
+import {
+  FeatureAggregationData,
+  FeatureAttributes,
+} from 'public/models/interfaces';
 import { darkModeEnabled } from '../../../../utils/kibanaUtils';
+import { prepareDataForChart } from '../../../DetectorResults/utils/anomalyResultUtils';
+// import { CodeModal } from '../../../DetectorConfig/components/CodeModal/CodeModal';
 
 interface FeatureChartProps {
-  title: string;
-  enabled: boolean;
+  feature: FeatureAttributes;
   onEdit(ev: React.MouseEvent<HTMLButtonElement>): void;
   isEdit: boolean;
   featureData: FeatureAggregationData[];
@@ -39,40 +44,83 @@ interface FeatureChartProps {
   isLoading: boolean;
   startDateTime: Moment;
   endDateTime: Moment;
+  featureType: string;
+  field?: string;
+  aggregationMethod?: string;
+  aggregationQuery?: string;
+  featureDataSeriesName: string;
+  featureAnomalyAnnotationSeriesName: string;
+  showAnomalyAsBar: boolean;
 }
 const getDisabledChartBackground = () =>
   darkModeEnabled() ? '#25262E' : '#F0F0F0';
 
 export const FeatureChart = (props: FeatureChartProps) => {
+  const [showCustomExpression, setShowCustomExpression] = useState<boolean>(
+    false
+  );
   const timeFormatter = niceTimeFormatter([
     props.startDateTime.valueOf(),
     props.endDateTime.valueOf(),
   ]);
   const showLoader = useDelayedLoader(props.isLoading);
 
+  const featureDescription = () => (
+    <EuiText size="s">
+      {props.featureType === 'simple_aggs' ? (
+        <p style={{ fontSize: '12px', color: 'grey' }}>
+          Field: {props.field}; Aggregation method: {props.aggregationMethod};
+          State: {props.feature.featureEnabled ? 'Enabled' : 'Disabled'}
+        </p>
+      ) : (
+        <p style={{ fontSize: '12px', color: 'grey' }}>
+          Custom expression:{' '}
+          <EuiLink onClick={() => setShowCustomExpression(true)}>
+            View code
+          </EuiLink>
+          ; State: {props.feature.featureEnabled ? 'Enabled' : 'Disabled'}{' '}
+        </p>
+      )}
+    </EuiText>
+  );
+  const customTheme = {
+    lineSeriesStyle: {
+      line: {
+        strokeWidth: 2,
+        visible: true,
+        opacity: 0.5,
+      },
+      point: {
+        visible: true,
+        stroke: '#16191F',
+      },
+    },
+  };
+
+  const featureData = prepareDataForChart(
+    props.featureData,
+    props.startDateTime,
+    props.endDateTime
+  );
   return (
     <ContentPanel
-      title={props.enabled ? props.title : `${props.title} ( disabled )`}
-      titleSize="xs"
-      titleClassName="preview-title"
+      title={
+        props.feature.featureEnabled
+          ? props.feature.featureName
+          : `${props.feature.featureName} ( disabled )`
+      }
       panelStyles={props.isEdit ? { border: '5px solid #96C8DA' } : {}}
       bodyStyles={
-        !props.enabled ? { backgroundColor: getDisabledChartBackground() } : {}
+        !props.feature.featureEnabled
+          ? { backgroundColor: getDisabledChartBackground() }
+          : {}
       }
-      actions={
-        <EuiButton
-          onClick={props.onEdit}
-          disabled={props.isEdit}
-          data-test-subj="editFeature"
-        >
-          Edit
-        </EuiButton>
-      }
+      subTitle={featureDescription()}
     >
       {props.featureData.length > 0 ? (
         <div
           style={{
-            height: '300px',
+            height: '200px',
             width: '100%',
             opacity: showLoader ? 0.2 : 1,
           }}
@@ -80,40 +128,55 @@ export const FeatureChart = (props: FeatureChartProps) => {
           <Chart>
             <Settings
               showLegend
-              legendPosition="bottom"
+              legendPosition={Position.Right}
+              theme={[customTheme]}
             />
-            {props.enabled ? (
+            {props.feature.featureEnabled ? (
               <RectAnnotation
                 dataValues={props.annotations || []}
                 id="annotations"
-                // annotationId={getAnnotationId('react')}
                 style={{
-                  stroke: darkModeEnabled() ? 'red' : '#FCAAAA',
+                  stroke: darkModeEnabled() ? 'red' : '#D5DBDB',
                   strokeWidth: 1,
                   opacity: 0.8,
-                  fill: darkModeEnabled() ? 'red' : '#FCAAAA',
+                  fill: darkModeEnabled() ? 'red' : '#D5DBDB',
                 }}
               />
             ) : null}
-            <Axis id="left" title={props.title} position="left" />
+            <Axis
+              id="left"
+              title={props.featureDataSeriesName}
+              position="left"
+              showGridLines
+            />
             <Axis id="bottom" position="bottom" tickFormat={timeFormatter} />
             <LineSeries
-              id="lines"
-              name={`Aggregated data for ${props.title}`}
+              id="featureData"
+              name={props.featureDataSeriesName}
               xScaleType="time"
               yScaleType="linear"
-              xAccessor={'startTime'}
+              xAccessor={'plotTime'}
               yAccessors={['data']}
-              data={props.featureData}
+              color={['#16191F']}
+              data={featureData}
             />
           </Chart>
+          {/* {showCustomExpression ? (
+            <CodeModal
+              title={props.feature.featureName}
+              subtitle="Custom expression"
+              code={JSON.stringify(props.feature.aggregationQuery, null, 4)}
+              getModalVisibilityChange={() => true}
+              closeModal={() => setShowCustomExpression(false)}
+            />
+          ) : null} */}
         </div>
       ) : (
         <EuiEmptyPrompt
           style={{ maxWidth: '45em' }}
           body={
             <EuiText>
-              <p>{`There is no data to display for feature ${props.title}`}</p>
+              <p>{`There is no data to display for feature ${props.feature.featureName}`}</p>
             </EuiText>
           }
         />
